Refactor int to int64 where used as data

+++ pkg/todoserver/models.go
@@ -2,7 +2,7 @@ package todoserver
 
 // TodoEntry is a todo by a user
 type TodoEntry struct {
-	ID      int    `json:"id"`
+	ID      int64  `json:"id"`
 	Title   string `json:"title"`
-	OwnerID int    `json:"ownerID"`
+	OwnerID int64  `json:"ownerID"`
 }
