+++ pkg/todoserver/todo-repository.go
@@ -2,8 +2,8 @@ package todoserver
 
 // TodoRepository is an interface for retrieving data
 type TodoRepository interface {
-	GetTodoByID(id int) *TodoEntry
-	GetTodosForUser(userID int) []TodoEntry
+	GetTodoByID(id int64) *TodoEntry
+	GetTodosForUser(userID int64) []TodoEntry
 	InsertTodo(todo TodoEntry) (*TodoEntry, error)
-	DeleteTodo(id int) error
+	DeleteTodo(id int64) error
 }
